Move form mixins from extras app to netbox app
