+++ netbox/core/forms/filtersets.py
@@ -4,8 +4,8 @@ from django.utils.translation import gettext_lazy as _
 
 from core.choices import *
 from core.models import *
-from extras.forms.mixins import SavedFiltersMixin
 from netbox.forms import NetBoxModelFilterSetForm
+from netbox.forms.mixins import SavedFiltersMixin
 from netbox.utils import get_data_backend_choices
 from utilities.forms import BOOLEAN_WITH_BLANK_CHOICES, FilterForm
 from utilities.forms.fields import ContentTypeChoiceField, DynamicModelMultipleChoiceField
