+++ netbox/dcim/forms/bulk_create.py
@@ -1,9 +1,9 @@
 from django import forms
+from django.utils.translation import gettext_lazy as _
 
 from dcim.models import *
-from django.utils.translation import gettext_lazy as _
-from extras.forms import CustomFieldsMixin
 from extras.models import Tag
+from netbox.forms.mixins import CustomFieldsMixin
 from utilities.forms import BootstrapMixin, form_from_model
 from utilities.forms.fields import DynamicModelMultipleChoiceField, ExpandableNameField
 from .object_create import ComponentCreateForm
