+++ netbox/extras/forms/__init__.py
@@ -3,5 +3,4 @@ from .filtersets import *
 from .bulk_edit import *
 from .bulk_import import *
 from .misc import *
-from .mixins import *
 from .scripts import *
